+++ deezy/utils/utils.py
@@ -1,6 +1,6 @@
+from pathlib import Path
 import re
 import sys
-from pathlib import Path
 
 from unidecode import unidecode